requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scikit-learn==1.3.2
joblib==1.3.2
websocket-client==1.7.0
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
matplotlib==3.8.2
schedule==1.2.1
Flask==3.0.0
flask-cors==4.0.0
